Add the v1.0.0 Mixpanel tag to the repo
The other 3 spec versions for this library seem to point to arbitrary commits
in the mixpanel repo. The spec I have created points to the v1.0.0 tag.
This tag's api seems to match the mixpanel instructions leading me to
believe this tag is the commit Mixpanel is expecting user to use. I
think spec versions 0.0.1, 1.2.1 and 1.3.1 should be removed.