[New Pod] OpenSSL
OpenSSL is deprecated in Mac OS 10.7, and gone on iOS. It is however
used by many open source projects, and there's nothing wrong with it
except Apple doesn't want to maintain API compatibility for it.

So if you want to use it in a non-deprecated manner, just compile
it into your project with this pod.